+++ kibot/kiplot.py
@@ -119,6 +119,16 @@ def load_actions(progress=None):
         activate.deactivate()
 
 
+def filter_warning(text):
+    if not GS.ci_cd_detected:
+        return False
+    if 'Missing KiCad main config file' in text:
+        return True
+    if 'Missing default system table' in text:
+        return True
+    return False
+
+
 def extract_errors(text):
     in_error = in_warning = False
     msg = ''
@@ -132,7 +142,10 @@ def extract_errors(text):
                 logger.error(msg.rstrip())
             elif in_warning:
                 in_warning = False
-                logger.warning(W_KIAUTO+msg.rstrip())
+                msg = W_KIAUTO+msg.rstrip()
+                if filter_warning(msg):
+                    msg = W_SILLY+msg
+                logger.warning(msg)
         if line.startswith('ERROR:'):
             in_error = True
             msg = line[6:]
@@ -144,7 +157,10 @@ def extract_errors(text):
         logger.error(msg.rstrip())
     elif in_warning:
         in_warning = False
-        logger.warning(W_KIAUTO+msg.rstrip())
+        msg = W_KIAUTO+msg.rstrip()
+        if filter_warning(msg):
+            msg = W_SILLY+msg
+        logger.warning(msg)
 
 
 def debug_output(res):